Bug fixed
The user was able to open three cards by quickly clicking on them successsively. This was making the application buggy as at one particular time no more than two cards should be open. So whenever a card is performing it's open animation the press events to all other cards are ignored until that animation completes. This ensures the desired result.